requirements.txt
uvicorn[standard]
fastapi
pydantic[Email]
python-dotenv
passlib
bcrypt
python-jose[cryptography]
python-multipart
httpx
pytest
fastapi-limiter
beanie
pillow
APScheduler
pytz
cloudinary
motor
jwt